made each common character have different passiveincome and passiveincometime
made it so that the characters passivetime the longer it is the player will get more at least to compensate. this will also add the random passive income when the player sells the unit and passive income.

relearning the whole system again too along with thinking about adding my racing game from itch.io into this game