user can now change the characters name & tested making a CHAO character
set up a system where the user is now able to change the name of a character and it'll also save their names for it. also had to fix some issues with it where a user while typing could close the menu or change the character as they were typing so had to lock them in a typing state. also learned about input fields and was able to get this system going. probably better ways to do this but this is at least working for now.

on top of this also created a test CHAO character just to see what the flow is like creating new types we'll see what i eventually end up doing but its not the worst for now. might create a bowser one or smth next time.