made a bowser/gucci/minecraft prefabs, changed purchase system for multiple
managed to set up the buy system so that the user is able to actually have odds of getting different versions now.

and for now creating characters the best way is to just create prefabs at the moment. changed the gamemanager system to now detect the tags of the tier of character to do different things in the testA script too. wanted to do different scripts for different characters but for now this is just the better method.

planning on buffing the moneyTimer for the legendaryTier along with having to polish up some more things too with the prefabs. also gotta find a way to make the characters more visible when colored versions